[circuit-writer] factor out writer mod from lib and expose prologue
